Share AD test bodies across backends via test/shared/
Each AD test body lives in one place under `test/shared/` (helpers, bijector
test functions, and vector-distribution lists with parameterised `test_all`
wrappers). The main suite and the Enzyme integration test each pick their
adtype lists and call the same shared helpers, eliminating ~550 lines of
near-duplicate test code across `test/ad/*.jl`, `test/vector/*.jl`, and
`test/integration/enzyme/main.jl`.
