+++ test/ad/pd.jl
@@ -1,21 +1,3 @@
-_topd(x) = x * x' + I
-
 @testset "PDVecBijector: $backend_name" for (backend_name, adtype) in TEST_ADTYPES
-    d = 4
-    b = Bijectors.PDVecBijector()
-    binv = inverse(b)
-
-    z = randn(d, d)
-    x = _topd(z)
-    y = b(x)
-
-    forward_only(x) = sum(transform(b, _topd(reshape(x, d, d))))
-    inverse_only(y) = sum(transform(binv, y))
-    inverse_chol_lower(y) = sum(Bijectors.cholesky_lower(transform(binv, y)))
-    inverse_chol_upper(y) = sum(Bijectors.cholesky_upper(transform(binv, y)))
-
-    test_ad(forward_only, adtype, vec(z))
-    test_ad(inverse_only, adtype, y)
-    test_ad(inverse_chol_lower, adtype, y)
-    test_ad(inverse_chol_upper, adtype, y)
+    test_pdvecbijector_ad(adtype)
 end
