+++ test/runtests.jl
@@ -39,14 +39,10 @@ const TEST_ADTYPES = [
     ("ReverseDiffCompiled", AutoReverseDiff(; compile=true)),
     ("Mooncake", AutoMooncake()),
 ]
-const REF_BACKEND = AutoFiniteDifferences(; fdm=central_fdm(5, 1))
 
-function test_ad(f, backend, x; rtol=1e-6, atol=1e-6)
-    @info "testing AD for function $f with $backend"
-    ref_gradient = DifferentiationInterface.gradient(f, REF_BACKEND, x)
-    gradient = DifferentiationInterface.gradient(f, backend, x)
-    @test isapprox(gradient, ref_gradient; rtol=rtol, atol=atol)
-end
+include("shared/ad_test_utils.jl")
+include("shared/ad_bijector_tests.jl")
+include("shared/vector_distributions.jl")
 
 # Always include this since it can be useful for other tests.
 include("bijectors/utils.jl")
