+++ test/vector/matrix.jl
@@ -1,56 +1,27 @@
 module VBMatrixTests
 
-using Distributions
-using LinearAlgebra
-using Test
-using PDMats
 import DifferentiationInterface as DI
-using Bijectors.VectorBijectors
 using ForwardDiff: ForwardDiff
-using ReverseDiff: ReverseDiff
 using Mooncake: Mooncake
+using ReverseDiff: ReverseDiff
+using Test
 
-ν = 5
-M = [1 2 3; 4 5 6]
-Σ = PDMats.PDMat([1 0.5; 0.5 1])
-Ω = PDMats.PDMat([1 0.3 0.2; 0.3 1 0.4; 0.2 0.4 1])
+include(joinpath(@__DIR__, "..", "shared", "vector_distributions.jl"))
 
-# TODO(penelopeysm): ReverseDiff gives wrong results when differentiating
-# through VecCorrBijector. Correctness tests are disabled for now.
-# https://github.com/TuringLang/Bijectors.jl/issues/434
-lkj_test_adtypes = [DI.AutoMooncake(), DI.AutoMooncakeForward()]
-
-# Don't check that from_linked_vec(d)(randn(...)) is in support for LKJ,
-# The reason is because the inverse bijector for LKJ causes the diagonal
-# entries to be not exactly 1 due to numerical precision issues. This
-# should in principle be fixed, but for now we skip the test.
-# https://github.com/TuringLang/Bijectors.jl/issues/435
-test_in_support(d) = !(d isa LKJ)
-
-matrix_dists = [
-    MatrixNormal(2, 4),
-    MatrixNormal(3, 5),
-    MatrixTDist(ν, M, Σ, Ω),
-    Wishart(7, Matrix{Float64}(I, 2, 2)),
-    Wishart(7, Matrix{Float64}(I, 4, 4)),
-    InverseWishart(7, Matrix{Float64}(I, 2, 2)),
-    InverseWishart(7, Matrix{Float64}(I, 4, 4)),
-    # See note in matrix/posdef.jl
-    # MatrixBeta(3, 3, 1000),
-    # MatrixBeta(5, 8, 1000),
-    LKJ(3, 1.0),
-    LKJ(7, 1.0),
+# Enzyme is tested separately in test/integration/enzyme.
+const adtypes = [
+    DI.AutoReverseDiff(),
+    DI.AutoReverseDiff(; compile=true),
+    DI.AutoMooncake(),
+    DI.AutoMooncakeForward(),
 ]
 
+# ReverseDiff gives wrong results when differentiating through VecCorrBijector, so we run
+# LKJ with Mooncake only. https://github.com/TuringLang/Bijectors.jl/issues/434
+const lkj_adtypes = [DI.AutoMooncake(), DI.AutoMooncakeForward()]
+
 @testset "Matrix distributions" begin
-    for d in matrix_dists
-        kwargs = (expected_zero_allocs=(), test_in_support=test_in_support(d))
-        if d isa LKJ
-            VectorBijectors.test_all(d; adtypes=lkj_test_adtypes, kwargs...)
-        else
-            VectorBijectors.test_all(d; kwargs...)
-        end
-    end
+    test_matrix_dists_with(adtypes; lkj_adtypes=lkj_adtypes)
 end
 
 end # module VBMatrixTests
