+++ test/vector/product.jl
@@ -1,105 +1,23 @@
 module VBProductTests
 
-using Distributions
-using LinearAlgebra
-using FillArrays: Fill
-using Test
-using Bijectors.VectorBijectors
 import DifferentiationInterface as DI
 using ForwardDiff: ForwardDiff
-using ReverseDiff: ReverseDiff
 using Mooncake: Mooncake
+using ReverseDiff: ReverseDiff
+using Test
 
-adtypes = [
+include(joinpath(@__DIR__, "..", "shared", "vector_distributions.jl"))
+
+# Enzyme is tested separately in test/integration/enzyme.
+const adtypes = [
     DI.AutoReverseDiff(),
     DI.AutoReverseDiff(; compile=true),
     DI.AutoMooncake(),
     DI.AutoMooncakeForward(),
 ]
 
-# These are purposely chosen because the vec_length output is the same but
-# linked_vec_length differs.
-m2 = MvNormal(zeros(2), I)
-d2 = Dirichlet(ones(2))
-
-p1t = product_distribution(Normal(), Beta(2, 2))
-p2t = product_distribution(m2, d2)
-p1a = product_distribution(fill(Beta(2, 2), 2))
-p2a = product_distribution(fill(d2, 2))
-
-products = [
-    # Tuples
-    product_distribution(Normal()),
-    product_distribution(Normal(), Normal()),
-    product_distribution(Normal(), Beta(2, 2)),
-    product_distribution(Beta(2, 2), Exponential()),
-    product_distribution(m2, d2),
-    product_distribution(m2, d2, m2, d2),
-    # Vectors of univariate (Distributions.Product)
-    product_distribution(fill(Normal(), 2)), # This is actually an MvNormal in disguise
-    product_distribution(fill(Beta(2, 2), 2)),
-    product_distribution([Uniform(0, 1), Uniform(1, 2), Uniform(2, 3)]),
-    product_distribution(Fill(Uniform(1, 2), 2)),
-    # >1D arrays, or vectors of >1D distributions (Distributions.ProductDistribution)
-    product_distribution(fill(Normal(), 2, 2)),
-    product_distribution(Fill(Uniform(1, 2), 2, 2)),
-    product_distribution(fill(m2, 2, 2)),
-    product_distribution(Fill(m2, 2, 2)),
-    product_distribution(fill(d2, 2, 2)),
-    # NamedTuples
-    product_distribution((a=Normal(), b=Beta(2, 2))),
-    product_distribution((a=Normal(), b=Dirichlet(ones(2)))),
-    product_distribution((a=Normal(), b=product_distribution(fill(Beta(2, 2), 2)))),
-    # Nested
-    product_distribution(fill(p1t, 2)),
-    product_distribution(fill(p1t, 2, 2)),
-    product_distribution(p2t, p2t, p2t),
-    product_distribution(fill(p2t, 2)),
-    product_distribution(fill(p2t, 2, 2)),
-    product_distribution(fill(p1a, 2)),
-    product_distribution(fill(p1a, 2, 2)),
-    product_distribution(p2a, p2a, p2a),
-    product_distribution(fill(p2a, 2)),
-    product_distribution(fill(p2a, 2, 2)),
-]
-
-nested_product_namedtuple = [
-    # This one is very weird: on Julia 1.10 (and only on Julia 1.10), @inferred to_vec(d)
-    # fails, but @code_warntype to_vec(d) is completely type stable. Almost certainly a
-    # Julia bug or at least a weird quirk
-    product_distribution((a=Normal(), b=product_distribution((c=Normal(), d=Beta(2, 2))))),
-]
-
-type_unstable_products = [
-    # Heterogeneous arrays make bijector construction type unstable. The triple-nested tuple
-    # products were also originally grouped here because Enzyme could not differentiate
-    # through them; Enzyme is now tested separately in test/integration/enzyme.
-    product_distribution([Normal(), Beta(2, 2), Exponential()]),
-    product_distribution([Normal() Beta(2, 2); Exponential() Uniform(-1, 1)]),
-    product_distribution([m2 d2; m2 d2]),
-    product_distribution(p1t, p1t, p1t),
-    product_distribution(p1a, p1a, p1a),
-]
-
 @testset "Product distributions" begin
-    for d in products
-        VectorBijectors.test_all(d; adtypes=adtypes, expected_zero_allocs=())
-    end
-
-    for d in nested_product_namedtuple
-        VectorBijectors.test_all(
-            d;
-            adtypes=adtypes,
-            expected_zero_allocs=(),
-            test_construction_type_stable=(VERSION >= v"1.11-"),
-        )
-    end
-
-    for d in type_unstable_products
-        VectorBijectors.test_all(
-            d; adtypes=adtypes, expected_zero_allocs=(), test_construction_type_stable=false
-        )
-    end
+    test_products_with(adtypes)
 end
 
 end # module VBProductTests
