+++ test/vector/reshaped.jl
@@ -1,37 +1,23 @@
 module VBReshapedTests
 
-using Distributions
-using LinearAlgebra
-using Test
-using Bijectors.VectorBijectors
 import DifferentiationInterface as DI
 using ForwardDiff: ForwardDiff
-using ReverseDiff: ReverseDiff
 using Mooncake: Mooncake
+using ReverseDiff: ReverseDiff
+using Test
 
-reshaped = [
-    # 0-dim array output: doesn't work because
-    # https://github.com/JuliaStats/Distributions.jl/issues/2025
-    # reshape(Normal(), ()),
-    vec(Normal()),
-    reshape(Normal(), (1, 1, 1, 1, 1)),
-    vec(Beta(2, 2)),
-    vec(Poisson(3)),
-    reshape(Poisson(3), (1, 1, 1, 1, 1)),
-    reshape(MvNormal(zeros(2), I), (2, 1, 1)),
-    reshape(MvNormal(zeros(4), I), (2, 2)),
-    reshape(Dirichlet(ones(6)), (2, 3)),
-    reshape(MatrixNormal(2, 4), 8),
-    reshape(MatrixNormal(2, 5), 5, 2),
-    reshape(Wishart(7, Matrix{Float64}(I, 4, 4)), 16),
-    reshape(Wishart(7, Matrix{Float64}(I, 4, 4)), 1, 1, 4, 1, 4),
-    reshape(Beta(2, 2), (1, 1, 1, 1, 1)),
+include(joinpath(@__DIR__, "..", "shared", "vector_distributions.jl"))
+
+# Enzyme is tested separately in test/integration/enzyme.
+const adtypes = [
+    DI.AutoReverseDiff(),
+    DI.AutoReverseDiff(; compile=true),
+    DI.AutoMooncake(),
+    DI.AutoMooncakeForward(),
 ]
 
 @testset "Reshaped distributions" begin
-    for d in reshaped
-        VectorBijectors.test_all(d; expected_zero_allocs=())
-    end
+    test_reshaped_with(adtypes)
 end
 
 end # module VBReshapedTests
